Fixed Transaction_ID error for ModbusTCP and type conversion issues for protocol resolution.

+++ addregdialog.h
@@ -15,7 +15,7 @@ struct ModbusRegReadDefinitions
     quint8 id;
     quint8 function;
     quint16 reg_addr;
-    quint8 quantity;
+    quint32 quantity;
     quint32 scan_rate;
     QByteArray packet;
 };
