+++ modbus_ascii.cpp
@@ -85,7 +85,7 @@ ModbusFrameInfo Modbus_ASCII::masterPack2Frame(const QByteArray &pack)
     else if(ret.function == ModbusWriteMultipleCoils
              || ret.function == ModbusWriteMultipleRegisters)
     {
-        ret.quantity = hex_pack[4] << 8 | hex_pack[5];
+        ret.quantity = quint16(hex_pack[4]) << 8 | quint8(hex_pack[5]);
     }
     else if(ret.function > ModbusFunctionError)
     {
@@ -160,20 +160,20 @@ ModbusFrameInfo Modbus_ASCII::slavePack2Frame(const QByteArray &pack)
         ret.function == ModbusReadHoldingRegisters ||
         ret.function == ModbusReadInputRegisters)
     {
-        ret.reg_addr = hex_pack[2] << 8 | hex_pack[3];
-        ret.quantity = hex_pack[4] << 8 | hex_pack[5];
+        ret.reg_addr = quint16(hex_pack[2]) << 8 | quint8(hex_pack[3]);
+        ret.quantity = quint16(hex_pack[4]) << 8 | quint8(hex_pack[5]);
     }
     else if(ret.function == ModbusWriteSingleCoil ||
              ret.function == ModbusWriteSingleRegister)
     {
-        ret.reg_addr = hex_pack[2] << 8 | hex_pack[3];
+        ret.reg_addr = quint16(hex_pack[2]) << 8 | quint8(hex_pack[3]);
         ret.quantity = 1;
-        ret.reg_values[0] = hex_pack[4] << 8 | quint8(hex_pack[5]);
+        ret.reg_values[0] = quint16(hex_pack[4]) << 8 | quint8(hex_pack[5]);
     }
     else if(ret.function == ModbusWriteMultipleCoils)
     {
-        ret.reg_addr = hex_pack[2] << 8 | hex_pack[3];
-        ret.quantity = hex_pack[4] << 8 | hex_pack[5];
+        ret.reg_addr = quint16(hex_pack[2]) << 8 | quint8(hex_pack[3]);
+        ret.quantity = quint16(hex_pack[4]) << 8 | quint8(hex_pack[5]);
         int byte_num =  quint8(hex_pack[6]);
         unsigned char *coils = (unsigned char *)ret.reg_values;
         for(int i = 0;i < byte_num;++i)
@@ -183,8 +183,8 @@ ModbusFrameInfo Modbus_ASCII::slavePack2Frame(const QByteArray &pack)
     }
     else if(ret.function == ModbusWriteMultipleRegisters)
     {
-        ret.reg_addr = hex_pack[2] << 8 | hex_pack[3];
-        ret.quantity = hex_pack[4] << 8 | hex_pack[5];
+        ret.reg_addr = quint16(hex_pack[2]) << 8 | quint8(hex_pack[3]);
+        ret.quantity = quint16(hex_pack[4]) << 8 | quint8(hex_pack[5]);
         for(int i = 0;i < ret.quantity; ++i)
         {
             ret.reg_values[i] = quint16(hex_pack[7 + 2 * i]) << 8 | quint8(hex_pack[8 + i * 2]);
